+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 ### Fixed
 
+- Running with parrot input phase labels
 - Main module execution using -m
 
 ## [2.1.0] - 2022-01-03